+++ f4-math/include/f4/math/eigen_real.hpp
@@ -41,6 +41,10 @@
 #include <stdexcept>
 #include <vector>
 
+#ifndef M_PI
+#define M_PI 3.14159265358979323846
+#endif // !M_PI
+
 namespace f4::math {
 
 // ============================================================================